Complete skeletal layout of website
Overall skeletal layout of the website is there now, with items at where
they will be in the end result, just lacking adjustments to be proper,
aesthetic and matching the desired outcome. The dashboard was filled in
this time with dummy text and boxes, giving the website an extremely
rough overview of what it will look like in the end.

Next, work to finish the navigation bar completely before moving on to
the header and then the main content chunk, before finally setting the
overall design of the page.