Settle rest of design and layout
Finally put together the entire website after settling the layout of the
far right 2 containers and filling in content for everything, making the
page more realistic and less empty. Measures were also taken to make the
website more interactive and life-like through wrapping certain elements
in a tags and adding styles to simulate interactivity.

However, the website can still be improved on but I think my knowledge
is lacking to do the features I have in mind, such as making the project
divs clickable while still allowing the text to be highlighted from
within. It can also be improved a lot on scaling and stretching and
essentially flexibility and stretching of the divs and fonts to fit
different screen sizes and scales.